[CCUBE-1060][KT] Rename element-list-panel to main-panel to reduce confusion
