--- src/components/element-list-panel/index.ts
@@ -1,1 +1,0 @@
-export * from "./element-list-panel";
