+++ src/components/index.ts
@@ -1,4 +1,3 @@
 export * from "./common";
-export * from "./element-list-panel";
-export * from "./main";
+export * from "./main-panel";
 export * from "./side-panel";
