+++ src/components/main-panel/index.ts
@@ -1,0 +1,1 @@
+export * from "./main-panel";
