+++ src/components/main-panel/main-panel.styles.tsx
