+++ src/form-builder.tsx
@@ -1,13 +1,16 @@
-import { ElementListPanel, Main, SidePanel } from "./components";
+import { MainPanel, SidePanel } from "./components";
 import { BuilderProvider } from "./context-providers/builder";
+import { Container, Wrapper } from "./form-builder.styles";
 
 export const FormBuilder = () => {
     return (
         <BuilderProvider>
-            <Main>
-                <ElementListPanel />
-                <SidePanel />
-            </Main>
+            <Wrapper>
+                <Container type="grid" stretch>
+                    <MainPanel />
+                    <SidePanel />
+                </Container>
+            </Wrapper>
         </BuilderProvider>
     );
 };
